+++ lib/Differentiator/HessianModeVisitor.cpp
@@ -36,16 +36,14 @@ namespace clad {
   static const StringLiteral* CreateStringLiteral(ASTContext& C,
                                                   std::string str) {
     QualType CharTyConst = C.CharTy.withConst();
-    QualType StrTy =
-        clad_compat::getConstantArrayType(C, CharTyConst,
-                                          llvm::APInt(/*numBits=*/32,
-                                                      str.size() + 1),
-                                          /*SizeExpr=*/nullptr,
-                                          /*ASM=*/ArrayType::Normal,
-                                          /*IndexTypeQuals*/ 0);
-    const StringLiteral* SL =
-        StringLiteral::Create(C, str, /*Kind=*/clad_compat::StringKind_Ordinary,
-                              /*Pascal=*/false, StrTy, noLoc);
+    QualType StrTy = clad_compat::getConstantArrayType(
+        C, CharTyConst, llvm::APInt(/*numBits=*/32, str.size() + 1),
+        /*SizeExpr=*/nullptr,
+        /*ASM=*/clad_compat::ArraySizeModifier_Normal,
+        /*IndexTypeQuals*/ 0);
+    const StringLiteral* SL = StringLiteral::Create(
+        C, str, /*Kind=*/clad_compat::StringLiteralKind_Ordinary,
+        /*Pascal=*/false, StrTy, noLoc);
     return SL;
   }